refactor:重構login的className程式碼&將login放入page/login資料夾並改名為index
