+++ client/styles/WGS-login.module.css
@@ -1,4 +1,4 @@
-.WGSloginContainer{
+.WGS-loginContainer{
     height: 100vh;
     width: 100vw;
     overflow: hidden;
@@ -12,7 +12,7 @@
     padding: 0;
 }
 
-.WGSloginBgS{
+.WGS-loginBgS{
     position: absolute;
     right: 0;
     top: 10vh;
@@ -24,7 +24,7 @@
     text-wrap: nowrap;
 }
 
-.WGSloginBgT{
+.WGS-loginBgT{
     position: absolute;
     right: 0;
     top: 60vh;
@@ -37,7 +37,7 @@
 
 }
 
-.WGSloginCard {
+.WGS-loginCard {
     display: flex;
     width: 400px;
     height: 500px;
@@ -51,7 +51,7 @@
     box-shadow: 0px 0px 70px 8px rgba(255, 255, 255, 0.40);
     backdrop-filter: blur(28.450000762939453px);
 }
-.WGStitle{
+.WGS-title{
     color: #FFF;
     text-align: center;
     font-family: "DM Sans";
@@ -60,10 +60,10 @@
     font-weight: 400;
     line-height: normal;
 }
-.WGSloginForm{
+.WGS-loginForm{
     width: 100%;
 }
-.WGSloginInput{
+.WGS-loginInput{
     display: flex;
     height: 46px;
     width: 100%;
@@ -77,16 +77,16 @@
     border-bottom: 1px solid #FFF;
 }
 
-.WGSloginInput::placeholder{
+.WGS-loginInput::placeholder{
     color: rgba(255, 255, 255, 0.3);
 }
 
-.WGSpasswordContainer {
+.WGS-passwordContainer {
     margin-bottom: 10px;
     position: relative;
 }
 
-.WGSeyeIcon {
+.WGS-eyeIcon {
     position: absolute;
     right: 10px;
     top: 50%;
@@ -95,25 +95,25 @@
     cursor: pointer;
 }
 
-.WGSrememberMe {
+.WGS-rememberMe {
     display: flex;
     align-items: center;
     margin-bottom: 15px;
     color: white;
 }
 
-.WGSrememberMe input {
+.WGS-rememberMe input {
     margin-right: 10px;
     width: auto;
 }
 
-.WGSerrorMessage {
+.WGS-errorMessage {
     display: none;
     color: #ff6b6b;
     margin-bottom: 15px;
 }
 
-.WGSloginBtn{
+.WGS-loginBtn{
     display: flex;
     height: 46px;
     width: 100%;
@@ -127,14 +127,14 @@
     box-shadow: 0px 0px 10px 0px rgba(255, 255, 255, 0.25);
 }
 
-.WGSbottomLinks {
+.WGS-bottomLinks {
     display: flex;
     justify-content: space-between;
     align-items: center;
     align-self: stretch;
 }
 
-.WGSbottomLinks a {
+.WGS-bottomLinks a {
     color: white;
     text-decoration: none;
 }
